Migrate to Vite
