+++ packages/frontend/.env.development
@@ -1,5 +1,2 @@
-REACT_APP_API_URL="http://localhost:3000"
-REACT_APP_FRONTEND_URL="http://localhost:3001"
-BUILD_PATH='./dist'
-BROWSER='none'
-PORT=3001
+DDDFORUM_API_URL="http://localhost:3000"
+DDDFORUM_FRONTEND_URL="http://localhost:3001"
