+++ packages/frontend/.env.staging
@@ -1,3 +1,2 @@
-REACT_APP_API_URL="https://dddform-staging.onrender.com/"
-REACT_APP_FRONTEND_URL="http://localhost:3001"
-BUILD_PATH='./dist'
+DDDFORUM_API_URL="https://dddform-staging.onrender.com/"
+DDDFORUM_FRONTEND_URL="http://localhost:3001"
